+++ QuineMcCluskeyProject/src/hust/soict/ict/quinemccluskey/controller/outputcontroller/OutputController.java
@@ -2,9 +2,10 @@ package hust.soict.ict.quinemccluskey.controller.outputcontroller;
 
 
 import hust.soict.ict.quinemccluskey.controller.Controller;
-import hust.soict.ict.quinemccluskey.model.output.OutputFunction;
+import hust.soict.ict.quinemccluskey.model.output.CanonicalForm;
 import hust.soict.ict.quinemccluskey.model.table.IntermediateTable;
 import hust.soict.ict.quinemccluskey.model.table.PITable;
+import hust.soict.ict.quinemccluskey.utils.FunctionConverter;
 import javafx.fxml.FXML;
 import javafx.geometry.HPos;
 import javafx.geometry.VPos;
@@ -77,13 +78,13 @@ public class OutputController extends Controller {
     	}
     }
 
-    public void setMakeEquationTable(OutputFunction out) {
+    public void setMakeEquationTable(CanonicalForm out) {
     	for(int i = 0; i < out.getEPI().size(); i ++) {
     		Label EPILabel = new Label
     				(out.getEPI().get(i).getImplicant());
     		Label toBinaryLabel = new Label
     				(out.getEPI().get(i).getBinaryExpression());
-    		out.toCharacterEquation(out.getEPI().get(i));
+    		FunctionConverter.fromEPIToFunction(out.getEPI().get(i));
     		Label characterLabel = new Label
     				(out.getResult());
     		EPILabel.setFont(Font.font(EPILabel.getFont().getName(), 
@@ -111,7 +112,7 @@ public class OutputController extends Controller {
     	}
     }
     
-    public void setFinalEquation(OutputFunction out) {
+    public void setFinalEquation(CanonicalForm out) {
     	finalEquation.setText(out.getResult());
     }
 }
