+++ QuineMcCluskeyProject/src/hust/soict/ict/quinemccluskey/test/OutputFunctionTest.java
@@ -4,10 +4,9 @@ import java.util.ArrayList;
 
 import hust.soict.ict.quinemccluskey.model.Variable;
 import hust.soict.ict.quinemccluskey.model.column.Column;
-import hust.soict.ict.quinemccluskey.model.minterm.CombinedImplicant;
 import hust.soict.ict.quinemccluskey.model.minterm.Implicant;
 import hust.soict.ict.quinemccluskey.model.minterm.Minterm;
-import hust.soict.ict.quinemccluskey.model.output.OutputFunction;
+import hust.soict.ict.quinemccluskey.model.output.CanonicalForm;
 import hust.soict.ict.quinemccluskey.model.output.POS;
 import hust.soict.ict.quinemccluskey.model.output.SOP;
 import hust.soict.ict.quinemccluskey.model.table.IntermediateTable;
@@ -16,7 +15,7 @@ import hust.soict.ict.quinemccluskey.model.table.Table;
 
 public class OutputFunctionTest{
 	public static void main(String[] args) {
-		OutputFunction out = new SOP();
+		CanonicalForm out = new SOP();
 		ArrayList<Implicant> buff = new ArrayList<Implicant>();
 		Variable.numberDigits = 4;
